requirements.txt
absl-py==1.4.0
acvl-utils==0.2
addict==2.4.0
appdirs==1.4.4
asttokens==2.4.1
astunparse==1.6.3
autocvd==0.2.1
batchgenerators==0.25
bleach==6.0.0
bokeh==3.2.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.4
colorcet==3.0.1
colormath==3.0.0
comm==0.2.0
connected-components-3d==3.12.1
contourpy==1.1.0
cycler==0.11.0
dask==2023.7.1
datashader==0.15.1
datashape==0.5.2
debugpy==1.8.0
decorator==5.1.1
dicom2nifti==2.4.8
dlib==19.19.0
dnnlib @ https://github.com/podgorskiy/dnnlib/releases/download/0.0.1/dnnlib-0.0.1-py3-none-any.whl#sha256=e789afb6365198754f3eedd7687d2c321f6a2daf8009eb94e46fb8b90217361d
dnnlib-util==0.3.1
docker-pycreds==0.4.0
docopt==0.6.2
drive==0.4.4
dynamic-network-architectures==0.2
einops==0.6.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.37
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
graphviz==0.20.1
grpcio==1.56.2
h5py==3.9.0
holoviews==1.16.2
httplib2==0.20.4
huggingface-hub==0.16.4
idna==3.4
#imagecodecs==2023.7.4
imageio==2.31.1
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==5.12.0
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.1
jupyter_client==8.6.0
jupyter_core==5.5.0
keras==2.13.1
kiwisolver==1.4.4
kornia==0.6.12
lazy_loader==0.3
legacy==0.1.6
libclang==16.0.6
linecache2==1.0.0
linkify-it-py==2.0.2
lit==16.0.6
llvmlite==0.40.1
lmdb==1.4.1
locket==1.0.0
lpips==0.1.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mpmath==1.3.0
multipledispatch==1.0.0
nest-asyncio==1.5.8
networkx==3.1
nibabel==5.1.0
ninja==1.11.1
numba==0.57.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauth2client==4.1.3
oauthlib==3.2.2
openai-clip==1.0.1
opencv-python==4.8.0.74
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
panel==1.2.0
param==1.13.0
parso==0.8.3
partd==1.4.0
pathtools==0.1.2
pexpect==4.9.0
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.9.1
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyct==0.5.0
pydicom==2.4.1
Pygments==2.17.2
pyiqa==0.1.7
pynndescent==0.5.10
pyparsing==3.0.9
python-dateutil==2.8.2
python-gdcm==3.0.22
python-magic==0.4.27
pytorch-ignite==0.4.12
pytorch-msssim==1.0.0
pytz==2023.3
pyviz-comms==2.3.2
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.2
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
sentry-sdk==1.31.0
setproctitle==1.3.3
shapely==2.0.1
SimpleITK==2.2.1
six==1.16.0
skin-tone-classifier==0.2.2
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
tb-nightly==2.14.0a20230727
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.7.4
timm==0.9.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traceback2==1.4.0
traitlets==5.14.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.2
umap-learn==0.5.3
unittest2==1.1.0
uritemplate==4.1.1
urllib3==1.26.16
wandb==0.15.12
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.6
wrapt==1.15.0
xarray==2023.7.0
xyzservices==2023.7.0
yacs==0.1.8
yapf==0.40.1
yarg==0.1.9
zipp==3.15.0
